requirements.txt
fastapi
psycopg2-binary
python-dotenv
sqlalchemy
uvicorn